Refactor custom ListViewCell and TreeViewCell impl into separate classes
